+++ examples/DigitalIO.cpp
@@ -1,6 +1,5 @@
 #include <Arduino.h>
 
-#include <list>
 #include <memory>
 #include <string>
 #include <vector>
@@ -41,18 +40,18 @@ void setup() {
 
 		[] (Shell &shell __attribute__((unused)),
 				const std::vector<std::string> &arguments)
-				-> const std::list<std::string> {
+				-> const std::vector<std::string> {
 			if (arguments.size() == 1) {
 				/* The first argument has been provided, so return
 				 * completion values for the second argument.
 				 */
-				return std::list<std::string>{
+				return std::vector<std::string>{
 					read_flash_string(F("INPUT")),
 					read_flash_string(F("OUTPUT")),
 					read_flash_string(F("INPUT_PULLUP"))
 				};
 			} else {
-				return std::list<std::string>{};
+				return std::vector<std::string>{};
 			}
 		}
 	);
@@ -92,17 +91,17 @@ void setup() {
 
 		[] (Shell &shell __attribute__((unused)),
 				const std::vector<std::string> &arguments)
-				-> const std::list<std::string> {
+				-> const std::vector<std::string> {
 			if (arguments.size() == 1) {
 				/* The first argument has been provided, so return
 				 * completion values for the second argument.
 				 */
-				return std::list<std::string>{
+				return std::vector<std::string>{
 					read_flash_string(F("HIGH")),
 					read_flash_string(F("LOW"))
 				};
 			} else {
-				return std::list<std::string>{};
+				return std::vector<std::string>{};
 			}
 		}
 	);